+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.5] - 2024-02-22
+
 ### Added
 - export to dbt models (#37).
 - export to ODCS (#49).